app/requirements.txt
fastapi
uvicorn
SQLAlchemy
PyMySQL
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart
bcrypt==4.0.1
python-dotenv
alembic